+++ lib/td_ie_web/views/ingest_execution_view.ex
@@ -11,10 +11,14 @@ defmodule TdIeWeb.IngestExecutionView do
   end
 
   def render("ingest_execution.json", %{ingest_execution: ingest_execution}) do
-    %{id: ingest_execution.id,
+    %{
+      id: ingest_execution.id,
       ingest_id: ingest_execution.ingest_id,
       start_timestamp: ingest_execution.start_timestamp,
       end_timestamp: ingest_execution.end_timestamp,
-      status: ingest_execution.status}
+      status: ingest_execution.status,
+      file_name: ingest_execution.file_name,
+      file_size: ingest_execution.file_size
+    }
   end
 end